Updates of J-0

+++ app/src/main/java/com/android/esprit/smartreminders/activities/MainFrame.java
@@ -1,13 +1,7 @@
 package com.android.esprit.smartreminders.activities;
 
-import android.content.Context;
+import android.app.AlertDialog;
 import android.content.Intent;
-import android.content.IntentFilter;
-import android.hardware.Sensor;
-import android.hardware.SensorEvent;
-import android.hardware.SensorEventListener;
-import android.hardware.SensorManager;
-import android.net.wifi.WifiManager;
 import android.content.SharedPreferences;
 import android.os.Build;
 import android.os.Bundle;
@@ -22,16 +16,15 @@ import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.android.esprit.smartreminders.Entities.User;
-import com.android.esprit.smartreminders.Fragments.BrowseTasksFragment;
 import com.android.esprit.smartreminders.Fragments.EventsndMettingsFragment;
 import com.android.esprit.smartreminders.Fragments.FragmentChild;
 import com.android.esprit.smartreminders.Fragments.PlansFragment;
 import com.android.esprit.smartreminders.Fragments.ProfileFragment;
 import com.android.esprit.smartreminders.Fragments.ScheduleFragment;
 import com.android.esprit.smartreminders.Fragments.SettingsFragment;
-import com.android.esprit.smartreminders.Fragments.ShareTasksFragment;
 import com.android.esprit.smartreminders.Fragments.TasksFragment;
 import com.android.esprit.smartreminders.Fragments.ZonesFragment;
 import com.android.esprit.smartreminders.R;
@@ -41,6 +34,11 @@ import com.android.esprit.smartreminders.sessions.Session;
 public class MainFrame extends AppCommonsActivity implements NavigationView.OnNavigationItemSelectedListener {
 
     private User sessionUser;
+    private DrawerLayout drawer;
+    private Toolbar toolbar;
+    private ActionBarDrawerToggle toggle;
+    private NavigationView navigationView;
+
     private Class<? extends FragmentChild> visibleFragment;
 
     @Override
@@ -49,16 +47,16 @@ public class MainFrame extends AppCommonsActivity implements NavigationView.OnNa
         super.onCreate(savedInstanceState);
         clearBackStack();
         setContentView(R.layout.activity_main_frame);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
+        drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        toggle = new ActionBarDrawerToggle(
                 this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
         drawer.addDrawerListener(toggle);
         toggle.syncState();
 
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
+        navigationView = (NavigationView) findViewById(R.id.nav_view);
         navigationView.setNavigationItemSelectedListener(this);
         navigationView.setItemIconTintList(null);
         this.setTitle(R.string.app_name);
@@ -86,6 +84,7 @@ public class MainFrame extends AppCommonsActivity implements NavigationView.OnNa
             drawer.closeDrawer(GravityCompat.START);
         } else {
             super.onBackPressed();
+          //  updateHihlightedDrawer();
         }
 
     }
@@ -107,11 +106,11 @@ public class MainFrame extends AppCommonsActivity implements NavigationView.OnNa
         //noinspection SimplifiableIfStatement
         if (id == R.id.action_settings) {
             goToSettingsFragment();
-            this.visibleFragment=SettingsFragment.class;
+            this.visibleFragment = SettingsFragment.class;
             return true;
         } else if (id == R.id.action_profile) {
             goToProfileFragment();
-            this.visibleFragment=ProfileFragment.class;
+            this.visibleFragment = ProfileFragment.class;
             return true;
         }
         return super.onOptionsItemSelected(item);
@@ -123,7 +122,7 @@ public class MainFrame extends AppCommonsActivity implements NavigationView.OnNa
         // Handle navigation view item clicks here.
         int id = item.getItemId();
 
-         if (id == R.id.nav_schedule) {
+        if (id == R.id.nav_schedule) {
             goToScheduleFragment();
             visibleFragment = ScheduleFragment.class;
         } else if (id == R.id.nav_daily) {
@@ -135,12 +134,15 @@ public class MainFrame extends AppCommonsActivity implements NavigationView.OnNa
         } else if (id == R.id.nav_events) {
             goToEventndMeetingsFragment();
             visibleFragment = EventsndMettingsFragment.class;
-        }
-         else if (id == R.id.nav_zones) {
+        } else if (id == R.id.nav_zones) {
             goToZonesFragment();
             visibleFragment = ZonesFragment.class;
         } else if (id == R.id.action_logout) {
+
             logOut();
+        } else if (id == R.id.nav_settings) {
+            goToSettingsFragment();
+            visibleFragment = SettingsFragment.class;
         }
         DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
         drawer.closeDrawer(GravityCompat.START);
@@ -148,59 +150,46 @@ public class MainFrame extends AppCommonsActivity implements NavigationView.OnNa
     }
 
 
-
     private void goToScheduleFragment() {
         if (!ScheduleFragment.class.equals(visibleFragment))
-        switchFragmentsAddToBackStack(R.id.fragment_container, new ScheduleFragment());
+            switchFragmentsAddToBackStack(R.id.fragment_container, new ScheduleFragment());
 
     }
 
     private void goToPlansFragment() {
         if (!PlansFragment.class.equals(visibleFragment))
-        switchFragmentsAddToBackStack(R.id.fragment_container, new PlansFragment());
+            switchFragmentsAddToBackStack(R.id.fragment_container, new PlansFragment());
 
     }
 
     private void goToTasksFragment() {
         if (!TasksFragment.class.equals(visibleFragment))
-        switchFragmentsAddToBackStack(R.id.fragment_container, new TasksFragment());
+            switchFragmentsAddToBackStack(R.id.fragment_container, new TasksFragment());
 
     }
 
     private void goToEventndMeetingsFragment() {
         if (!EventsndMettingsFragment.class.equals(visibleFragment))
-        switchFragmentsAddToBackStack(R.id.fragment_container, new EventsndMettingsFragment());
-
-    }
-
-    private void goToShareTasksFragment() {
-        if (!ShareTasksFragment.class.equals(visibleFragment))
-        switchFragmentsAddToBackStack(R.id.fragment_container, new ShareTasksFragment());
-
-    }
-
-    private void goToBrowseTasksFragment() {
-        if (!BrowseTasksFragment.class.equals(visibleFragment))
-        switchFragmentsAddToBackStack(R.id.fragment_container, new BrowseTasksFragment());
+            switchFragmentsAddToBackStack(R.id.fragment_container, new EventsndMettingsFragment());
 
     }
 
     private void goToSettingsFragment() {
         if (!SettingsFragment.class.equals(visibleFragment))
-        switchFragmentsAddToBackStack(R.id.fragment_container, new SettingsFragment());
+            switchFragmentsAddToBackStack(R.id.fragment_container, new SettingsFragment());
 
     }
 
     private void goToProfileFragment() {
         if (!ProfileFragment.class.equals(visibleFragment))
-        switchFragmentsAddToBackStack(R.id.fragment_container, new ProfileFragment());
+            switchFragmentsAddToBackStack(R.id.fragment_container, new ProfileFragment());
 
     }
 
     public void goToUnStackedFragment(FragmentChild Child) {
-        if(!Child.getClass().equals(visibleFragment)) {
+        if (!Child.getClass().equals(visibleFragment)) {
 
-            visibleFragment=Child.getClass();
+            visibleFragment = Child.getClass();
             switchFragments(R.id.fragment_container, Child);
         }
 
@@ -208,16 +197,25 @@ public class MainFrame extends AppCommonsActivity implements NavigationView.OnNa
 
     public void goToZonesFragment() {
         if (!ZonesFragment.class.equals(visibleFragment))
-        switchFragments(R.id.fragment_container, new ZonesFragment());
+            switchFragments(R.id.fragment_container, new ZonesFragment());
         Log.d("Fragments Behavour", "goToZonesFragment: fragment changed !");
-
     }
 
     public void logOut() {
-        SharedPreferences sharedPref = getSharedPreferences("Myprefs", MODE_PRIVATE);
-        SharedPreferences.Editor editor = sharedPref.edit();
-        editor.clear().apply();
-        startActivity(new Intent(MainFrame.this, Login.class));
+        AlertDialog dialog = new AlertDialog.Builder(this, R.style.AlertDialogCustom)
+                .setIcon(R.drawable.ic_exit_to_app_black_24dp)
+                .setPositiveButton("Logout", (dialog1, which) -> {
+                    SharedPreferences sharedPref = getSharedPreferences("Myprefs", MODE_PRIVATE);
+                    SharedPreferences.Editor editor = sharedPref.edit();
+                    editor.clear().apply();
+                    startActivity(new Intent(MainFrame.this, Login.class));
+                })
+                .setNegativeButton("I changed my mind", (dialog1, which) -> {
+                    Toast.makeText(this, "Logout Canceled !", Toast.LENGTH_SHORT).show();
+                })
+                .setTitle("Logout").setMessage("Are you sure you want to Logout ? All our services will cease to function !  ").create();
+        dialog.show();
+
     }
 
     private void updateHihlightedDrawer() {
